==> 00-Packages_Imports_Directory.R <==

################################################
############## Working Directory ##############
################################################

setwd('C:\\Users\\up201806473\\Documents\\BA')

################################################
############# Imports and Packages #############
################################################

install.packages('caret')
install.packages('dplyr')
install.packages('data.table')
install.packages('robustbase')
install.packages("rcompanion")
install.packages("dataPreparation")
install.packages('forcats')
install.packages('naniar')
install.packages('moments')
install.packages('ltm')
install.packages('polycor')
install.packages('ggplot2')
install.packages('mltools')
install.packages('VIM')
install.packages('mice')
install.packages('Boruta')
install.packages('splitTools')
install.packages('performanceEstimation')
nstall.packages('csv')
install.packages("ggplot2")
install.packages('GGally')
install.packages('NbClust')
install.packages('neuralnet')
install.packages('caret')
install.packages('ROSE')
install.packages("ggplot2")
install.packages('GGally')
install.packages('NbClust')
install.packages('neuralnet')
install.packages('caret')
install.packages('ROSE')
install.packages('csv')
install.packages("writexl")
install.packages("splitTools")
install.packages("PresenceAbsence")
install.packages("randomForest")
install.packages("pROC")
install.packages('FNN')
install.packages('rattle')
install.packages('rpart.plot')
install.packages('e1071')
install.packages("pacman")
install.packages("lightgbm")
install.packages("ROCR")
install.packages("OneR")

library(OneR)
library(ROCR)
library(lightgbm)
library(pacman)
library(Matrix)
library (e1071)
library(lattice)
library(rattle)
library(rpart.plot)
library(rpart)
library(FNN)
library(splitTools)
library(PresenceAbsence)
library(pROC)
library(randomForest)
library(ROSE)
library(pROC)
library(naniar)       # NaN values study
library(caret)        #
library(dplyr)        # Data Manipulation
library(data.table)   # Extension of data.frame
library(robustbase)   # 
library(moments)      # contains the skewness function 
library(ltm)          # Since our target variable is dichotomous, we can use the point-biserial correlation. There is a function to do this in the ltm package
library(polycor)      # Polychoric correlation between ordinal variables
library(rcompanion)   # Cramer's V test between nominal variables
library(forcats)      # Combining factor levels
library(ggplot2)      #
library(mltools)      # Package that contains the one_hot function
library(VIM)          # package containing k-Nearest Neighbour Imputation
library(mice)         #
library(Boruta)       #
library(splitTools)   #Data partition for one of the datasets:
library(dataPreparation)        #data preparation package needed for target encoding
library(performanceEstimation)  #
library(NbClust)
library(GGally)
library(ggplot2)
library(csv)
library(neuralnet)
library(rpart)
library(pROC)
library(ROSE)
library(csv)
library(writexl)
